@@ -1,14 +1,14 @@
 plugins {
-    id("org.springframework.boot") version "3.5.4"
+    id("org.springframework.boot") version "3.5.7"
     id("io.spring.dependency-management") version "1.1.7"
     id("java")
-    id("com.github.jk1.dependency-license-report") version "2.9"
-    id("com.gorylenko.gradle-git-properties") version "2.5.2"
+    id("com.github.jk1.dependency-license-report") version "3.0.1"
+    id("com.gorylenko.gradle-git-properties") version "2.5.3"
 }
 
 group = "org.codeforamerica"
 version = "0.0.1-SNAPSHOT"
-java { sourceCompatibility = JavaVersion.VERSION_17 }
+java { sourceCompatibility = JavaVersion.VERSION_21 }
 
 tasks.withType(Jar){
     duplicatesStrategy = DuplicatesStrategy.EXCLUDE
