@@ -14,60 +14,60 @@ import org.junit.jupiter.api.Test;
 
 class HouseholdUsCitizenPreparerTest {
 
-  HouseholdUsCitizenPreparer preparer = new HouseholdUsCitizenPreparer();
+	 HouseholdUsCitizenPreparer preparer = new HouseholdUsCitizenPreparer();
 
-  @Test
-  void shouldParseOffWhenUsCitizenshipNotAsked() {
-    ApplicationData applicationData = new ApplicationData();
-    applicationData.getSubworkflows().addIteration("household", new PagesData());
-    UUID householdMemberID = applicationData.getSubworkflows().get("household").get(0).getId();
+	  @Test
+	  void shouldParseOffWhenUsCitizenshipNotAsked() {
+	    ApplicationData applicationData = new ApplicationData();
+	    applicationData.getSubworkflows().addIteration("household", new PagesData());
+	
 
-    List<DocumentField> result = preparer.prepareDocumentFields(Application.builder()
-        .applicationData(applicationData)
-        .build(), null, null);
+	    List<DocumentField> result = preparer.prepareDocumentFields(Application.builder()
+	        .applicationData(applicationData)
+	        .build(), null, null);
 
-    assertThat(result).isEqualTo(List.of(
-        new DocumentField(
-            "usCitizen",
-            "applicantIsUsCitizen",
-            List.of("Off"),
-            DocumentFieldType.SINGLE_VALUE,
-            null
-        )
-    ));
-  }
-    
-  @Test
-  void shouldParseFalseWhenApplicantIsNotUsCitizen() {
-      ApplicationData applicationData = new ApplicationData();
-      applicationData.getSubworkflows().addIteration("household", new PagesData());
-  UUID householdMemberID = applicationData.getSubworkflows().get("household").get(0).getId();
+	    assertThat(result).isEqualTo(List.of(
+	        new DocumentField(
+	            "usCitizen",
+	            "applicantIsUsCitizen",
+	            List.of("Off"),
+	            DocumentFieldType.SINGLE_VALUE,
+	            null
+	        )
+	    ));
+	  }
+	    
+	 @Test
+	  void shouldParseFalseWhenApplicantIsNotUsCitizen() {
+	      ApplicationData applicationData = new ApplicationData();
+	      applicationData.getSubworkflows().addIteration("household", new PagesData());
+	  UUID householdMemberID = applicationData.getSubworkflows().get("household").get(0).getId();
 
-  new TestApplicationDataBuilder(applicationData)
-  	  .withPageData("usCitizen", "isUsCitizen", "false")
-      .withPageData("whoIsNonCitizen", "whoIsNonCitizen",
-          List.of("personAFirstName personALastName applicant",
-              "personBFirstName personBLastName " + householdMemberID));
+	  new TestApplicationDataBuilder(applicationData)
+	  	  .withPageData("usCitizen", "isUsCitizen", "false")
+	      .withPageData("whoIsNonCitizen", "whoIsNonCitizen",
+	          List.of("personAFirstName personALastName applicant",
+	              "personBFirstName personBLastName " + householdMemberID));
 
-  List<DocumentField> result = preparer.prepareDocumentFields(Application.builder()
-      .applicationData(applicationData)
-      .build(), null, null);
+	  List<DocumentField> result = preparer.prepareDocumentFields(Application.builder()
+	      .applicationData(applicationData)
+	      .build(), null, null);
 
-  assertThat(result).isEqualTo(List.of(
-      new DocumentField(
-          "usCitizen",
-          "applicantIsUsCitizen",
-          List.of("false"),
-          DocumentFieldType.SINGLE_VALUE,
-          null
-      ),
-      new DocumentField(
-          "usCitizen",
-          "isUsCitizen",
-          List.of("false"),
-              DocumentFieldType.SINGLE_VALUE,
-              0
-          )
-      ));
-  }
-}
+	  assertThat(result).isEqualTo(List.of(
+	      new DocumentField(
+	          "usCitizen",
+	          "applicantIsUsCitizen",
+	          List.of("false"),
+	          DocumentFieldType.SINGLE_VALUE,
+	          null
+	      ),
+	      new DocumentField(
+	          "usCitizen",
+	          "isUsCitizen",
+	          List.of("false"),
+	              DocumentFieldType.SINGLE_VALUE,
+	              0
+	          )
+	      ));
+	  }
+	}
