@@ -38,7 +38,7 @@ class PageUtilsTest {
 		String[] namesArray = { "Julian Doyle applicant", "Aquila Graves d259c623-395e-4f7a-ba19-56b43447354b",
 				"Jayme Lamb 00942599-a681-41d2-8625-8a6538f560f7" };
 		List<String> namesList = Arrays.asList(namesArray);
-		Boolean isPresent = PageUtils.listOfNamesContainsName(namesList, "Aquila Graves");
+		Boolean isPresent = PageUtils.listOfNamesContainsName(namesList, "Aquila Graves d259c623-395e-4f7a-ba19-56b43447354b");
 		assertTrue(isPresent);
 	}
 
@@ -61,12 +61,31 @@ class PageUtilsTest {
 		List<String> result = Arrays.asList(childrenInNeedOfCare);
 		List<String> result2 = Arrays.asList(childrenInSchool);
 		List<String>result3= PageUtils.getEligibleSchoolAndChildCareMembers(result, result2);
-		Boolean isPresent = PageUtils.listOfNamesContainsName(result3, "Julian Doyle");
+		Boolean isPresent = PageUtils.listOfNamesContainsName(result3, "Julian Doyle applicant");
 		assertTrue(isPresent);
-		 isPresent = PageUtils.listOfNamesContainsName(result3, "Aquila Graves");
+		 isPresent = PageUtils.listOfNamesContainsName(result3, "Aquila Graves d259c623-395e-4f7a-ba19-56b43447354b");
 		assertTrue(isPresent);
-		 isPresent = PageUtils.listOfNamesContainsName(result3, "Jayme Lamb ");
+		 isPresent = PageUtils.listOfNamesContainsName(result3, "Jayme Lamb 00942599-a681-41d2-8625-8a6538f560f7");
 		assertTrue(isPresent);
+	}
+	
+	@Test
+	void shouldFindMemberWithSameNameInBothLists() {
+		String[] childrenInNeedOfCare = { "Julian Doyle applicant", "Julian Doyle d259c623-395e-4f7a-ba19-56b43447354b",
+				"Jayme Lamb 00942599-a681-41d2-8625-8a6538f560f7" };
+		String[] childrenInSchool = { "Julian Doyle applicant", "Julian Doyle d259c623-395e-4f7a-ba19-56b43447354b",
+		"Jayme Lamb 00942599-a681-41d2-8625-8a6538f560f7" };
+		List<String> result = Arrays.asList(childrenInNeedOfCare);
+		List<String> result2 = Arrays.asList(childrenInSchool);
+		List<String>result3= PageUtils.getEligibleSchoolAndChildCareMembers(result, result2);
+		Boolean isPresent = PageUtils.listOfNamesContainsName(result3, "Julian Doyle applicant");
+		assertTrue(isPresent);
+		 isPresent = PageUtils.listOfNamesContainsName(result3, "Julian Doyle d259c623-395e-4f7a-ba19-56b43447354b");
+		assertTrue(isPresent);
+		 isPresent = PageUtils.listOfNamesContainsName(result3, "Jayme Lamb 00942599-a681-41d2-8625-8a6538f560f7");
+		assertTrue(isPresent);
+		 isPresent = PageUtils.listOfNamesContainsName(result3, "Jayme Lamb 00942599-a681-41d2-8625");
+			assertFalse(isPresent);
 	}
 	
 	
