@@ -8,8 +8,6 @@ import java.util.List;
 import java.util.Map;
 
 import org.codeforamerica.shiba.application.FlowType;
-import org.codeforamerica.shiba.pages.config.FeatureFlag;
-import org.codeforamerica.shiba.pages.config.FeatureFlagConfiguration;
 import org.codeforamerica.shiba.pages.data.ApplicationData;
 import org.codeforamerica.shiba.testutilities.PagesDataBuilder;
 import org.codeforamerica.shiba.testutilities.TestApplicationDataBuilder;
@@ -17,10 +15,8 @@ import org.junit.jupiter.api.Test;
 
 public class WicRecommendationServiceTest {
 	
-	private FeatureFlagConfiguration featureFlagConfiguration = new FeatureFlagConfiguration(
-			Map.of("show-wic-recommendation", FeatureFlag.ON));
-	private WicRecommendationService wicRecommendationService = new WicRecommendationService(featureFlagConfiguration,
-			List.of("Anoka", "Carver"));
+
+	private WicRecommendationService wicRecommendationService = new WicRecommendationService(List.of("Anoka", "Carver"));
     ZonedDateTime now = ZonedDateTime.now();
     ZonedDateTime fiveYearsBefore = now.minusYears(5);    
 
